requirements.txt
# Core Machine Learning & Compute
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Graph Neural Networks
# Note: torch-scatter and torch-sparse often require specific binary installation 
# depending on CUDA version, but pip handles the base dependency.
torch-geometric>=2.3.0
dgl>=2.1.0  # CPU wheel available on PyPI; GPU users may prefer dgl-cu11 (updated for Python 3.12 compatibility)

# Quantum & ZX-Calculus
pyzx>=0.7.0      # Critical for converting Circuits <-> Graphs and checking gflow
qiskit>=1.0.0,<2.0.0      # For loading standard QASM benchmarks (e.g. MQTBench)
qiskit-aer                # Provides backend simulation

# Utilities
networkx>=3.0    # Graph data structures
tqdm>=4.65.0     # Progress bars
matplotlib>=3.7.0 # Visualization
mqt.bench
qiskit-ibm-runtime>=0.20.0  # Use V2 fake backends (FakeMontrealV2, etc.)
